[Chapter5] made http folder and move related java files
