+++ src/main/java/http/HttpRequest.java
@@ -1,7 +1,9 @@
-package util;
+package http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import util.HttpRequestUtils;
+import util.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
