+++ src/main/java/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package util;
+package http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
