+++ src/test/java/http/HttpRequestTest.java
@@ -1,5 +1,6 @@
-package util;
+package http;
 
+import http.HttpRequest;
 import org.junit.Test;
 
 import java.io.File;
