+++ src/test/java/http/HttpResponseTest.java
@@ -1,5 +1,6 @@
-package util;
+package http;
 
+import http.HttpResponse;
 import org.junit.Test;
 
 import java.io.File;
